Refactor ID definition and callSignature initialization mechanism of stubs
Refactor ID definition and callSignature initialization mechanism of runtime stub,
optimizer stub and bytecode handler stub,
reduce amount and complexity of macro used in calling those stubs.

Add stubDes struct to classify stubs both in AOT file generation phase
and AOT file loading phase.

Fix some circular dependency problems in including header files.

Fix certain inline function declaration and definition format problems, which could lead to
compiling errors.

Issue: https://gitee.com/openharmony/ark_js_runtime/issues/I4VMLD?from=project-issue
Test: stubTest(unit test), richards with asm interpreter enabled

+++ ecmascript/compiler/code_generator.h
@@ -25,7 +25,7 @@ class CodeGeneratorImpl {
 public:
     CodeGeneratorImpl() = default;
     virtual ~CodeGeneratorImpl() = default;
-    virtual void GenerateCodeForStub(Circuit *circuit, const ControlFlowGraph &graph, int index,
+    virtual void GenerateCodeForStub(Circuit *circuit, const ControlFlowGraph &graph, size_t index,
                                      const CompilationConfig *cfg) = 0;
 };
 
@@ -33,7 +33,7 @@ class CodeGenerator {
 public:
     explicit CodeGenerator(std::unique_ptr<CodeGeneratorImpl> &impl) : impl_(std::move(impl)) {}
     ~CodeGenerator() = default;
-    void Run(Circuit *circuit, const ControlFlowGraph &graph, int index, const CompilationConfig *cfg)
+    void Run(Circuit *circuit, const ControlFlowGraph &graph, size_t index, const CompilationConfig *cfg)
     {
         impl_->GenerateCodeForStub(circuit, graph, index, cfg);
     }
